+++ src/forms/form.rs
@@ -520,6 +520,7 @@ impl Form {
 }
 
 #[cfg(test)]
+#[allow(deprecated)]
 mod tests {
     use super::*;
     use crate::forms::{ConfirmField, InputField};